Monster and non-monster DAQmx now inherit tests
