+++ tests/DAQmx_monster.py
@@ -1,17 +1,11 @@
 import unittest
-from time import sleep
 
 from utils.test_modes import TestModes
-from utils.channel_access import ChannelAccess
 from utils.ioc_launcher import EPICS_TOP
-from utils.emulator_launcher import CommandLineEmulatorLauncher, DEVICE_EMULATOR_PATH, DAQMxEmulatorLauncher
-from utils.testing import get_running_lewis_and_ioc, assert_log_messages
+from utils.emulator_launcher import DAQMxEmulatorLauncher
+from common_tests.DAQmx import DAQmxTests, DEVICE_PREFIX
 
 import os
-
-
-# Device prefix
-DEVICE_PREFIX = "DAQMXTEST"
 
 IOCS = [
     {
@@ -32,29 +26,10 @@ IOCS = [
 TEST_MODES = [TestModes.DEVSIM]
 
 
-class DAQmxMonsterTests(unittest.TestCase):
+class DAQmxMonsterTests(DAQmxTests, unittest.TestCase):
     """
-    General tests for the DAQmx.
+    Test all DAQMx tests using monster mode.
     """
-    def setUp(self):
-        self.emulator, self._ioc = get_running_lewis_and_ioc(DEVICE_PREFIX, DEVICE_PREFIX)
-
-        self.ca = ChannelAccess(device_prefix=DEVICE_PREFIX)
-
-    def test_WHEN_emulator_disconnected_THEN_data_in_alarm_and_valid_on_reconnect(self):
-        self.ca.assert_that_pv_alarm_is_not("DATA", ChannelAccess.Alarms.INVALID)
-        self.emulator.disconnect_device()
-        self.ca.assert_that_pv_alarm_is("DATA", ChannelAccess.Alarms.INVALID)
-
-        # Check we don't get excessive numbers of messages if we stay disconnected for 15s (up to 15 messages seems
-        # reasonable - 1 per second on average)
-        with assert_log_messages(self._ioc, number_of_messages=15):
-            sleep(15)
-            # Double-check we are still in alarm
-            self.ca.assert_that_pv_alarm_is("DATA", ChannelAccess.Alarms.INVALID)
-
-        self.emulator.reconnect_device()
-        self.ca.assert_that_pv_alarm_is_not("DATA", ChannelAccess.Alarms.INVALID, timeout=5)
-        self.ca.assert_that_pv_value_is_changing("DATA", 1)
+    pass
